feat: consolidate suite_test file into common and functional_test files
